+++ src/commands/utility/help.js
@@ -1,5 +1,5 @@
 const { CommandCategory, BotClient } = require('@src/structures')
-const { EMBED_COLORS, SUPPORT_SERVER } = require('@root/config.js')
+const { EMBED_COLORS, SUPPORT_SERVER, DEV_IDS } = require('@root/config.js')
 const {
   EmbedBuilder,
   ActionRowBuilder,
@@ -46,7 +46,7 @@ module.exports = {
     if (!trigger) {
       const response = await getHelpMenu(message)
       const sentMsg = await message.safeReply(response)
-      return waiter(sentMsg, message.author.id, data.prefix)
+      return waiter(sentMsg, message.member, data.prefix)
     }
 
     // check if command help (!help cat)
@@ -67,7 +67,7 @@ module.exports = {
     if (!cmdName) {
       const response = await getHelpMenu(interaction)
       const sentMsg = await interaction.followUp(response)
-      return waiter(sentMsg, interaction.user.id)
+      return waiter(sentMsg, interaction.member)
     }
 
     // check if command help (!help cat)
@@ -85,11 +85,21 @@ module.exports = {
 /**
  * @param {CommandInteraction} interaction
  */
-async function getHelpMenu({ client, guild }) {
+async function getHelpMenu({ client, guild, member }) {
   // Menu Row
   const options = []
   for (const [k, v] of Object.entries(CommandCategory)) {
     if (v.enabled === false) continue
+    if (
+      (v.name.includes('Moderation') ||
+        v.name.includes('Admin') ||
+        v.name.includes('Automod') ||
+        v.name.includes('Giveaway')) &&
+      !member.permissions.has('ManageGuild')
+    ) {
+      continue
+    }
+    if (v.name === 'DEV' && !DEV_IDS.includes(member.user.id)) continue
     options.push({
       label: v.name,
       value: k,
@@ -141,13 +151,13 @@ async function getHelpMenu({ client, guild }) {
 
 /**
  * @param {Message} msg
- * @param {string} userId
+ * @param {import('discord.js').GuildMember} member
  * @param {string} prefix
  */
-const waiter = (msg, userId, prefix) => {
+const waiter = (msg, member, prefix) => {
   const collector = msg.channel.createMessageComponentCollector({
     filter: reactor =>
-      reactor.user.id === userId && msg.id === reactor.message.id,
+      reactor.user.id === member.id && msg.id === reactor.message.id,
     idle: IDLE_TIMEOUT * 1000,
     dispose: true,
     time: 5 * 60 * 1000,
@@ -167,7 +177,7 @@ const waiter = (msg, userId, prefix) => {
       case 'help-menu': {
         const cat = response.values[0].toUpperCase()
         arrEmbeds = prefix
-          ? getMsgCategoryEmbeds(msg.client, cat, prefix)
+          ? getMsgCategoryEmbeds(msg.client, cat, prefix, member)
           : getSlashCategoryEmbeds(msg.client, cat)
         currentPage = 0
 
@@ -320,7 +330,7 @@ function getSlashCategoryEmbeds(client, category) {
  * @param {string} category
  * @param {string} prefix
  */
-function getMsgCategoryEmbeds(client, category, prefix) {
+function getMsgCategoryEmbeds(client, category, prefix, member) {
   let collector = ''
 
   // For IMAGE Category
@@ -370,9 +380,11 @@ function getMsgCategoryEmbeds(client, category, prefix) {
       0,
       commands.length > CMDS_PER_PAGE ? CMDS_PER_PAGE : commands.length
     )
-    toAdd = toAdd.map(
-      cmd => `\`${prefix}${cmd.name}\`\n ❯ ${cmd.description}\n`
-    )
+    toAdd = toAdd
+      .filter(cmd =>
+        cmd.userPermissions ? member.permissions.has(cmd.userPermissions) : true
+      )
+      .map(cmd => `\`${prefix}${cmd.name}\`\n ❯ ${cmd.description}\n`)
     arrSplitted.push(toAdd)
   }