+++ src/helpers/Validator.js
@@ -69,7 +69,7 @@ module.exports = class Validator {
     }
 
     // Warnings
-    if (config.OWNER_IDS.length === 0) warn('config.js: OWNER_IDS are empty')
+    if (config.DEV_IDS.length === 0) warn('config.js: DEV_IDS are empty')
     if (!config.SUPPORT_SERVER)
       warn('config.js: SUPPORT_SERVER is not provided')
     if (!process.env.WEATHERSTACK_KEY)